# Optimize DataFrame.{head, tail} when DataFrame has unknown chunk shape