--- a/mars/optimizes/runtime/optimizers/cp.py
+++ b/mars/optimizes/runtime/optimizers/cp.py
@@ -30,9 +30,13 @@ CP_ELEMENTWISE_OP = {
 CP_OP = CP_ELEMENTWISE_OP
 
 
-class CpOptimizer(object):
+class CpRuntimeOptimizer:
     def __init__(self, graph):
         self._graph = graph
+
+    @classmethod
+    def is_available(cls):
+        return CP_INSTALLED
 
     def optimize(self, keys=None):
         self.compose(keys=keys)