--- a/mars/tensor/fuse/tests/test_cupy_execute.py
+++ b/mars/tensor/fuse/tests/test_cupy_execute.py
@@ -20,7 +20,7 @@ from mars.executor import Executor
 from mars.tensor.fuse.cp import _evaluate
 from mars.tensor.datasource import ones
 from mars.tensor import sqrt
-from mars.optimizes.runtime.optimizers.core import Optimizer
+from mars.optimizes.runtime.optimizers.core import RuntimeOptimizer
 
 
 class Test(unittest.TestCase):
@@ -33,7 +33,7 @@ class Test(unittest.TestCase):
         t = (t1 - t2) / sqrt(t2 * (1 - t2) * len(t2))
 
         g = t.build_graph(tiled=True)
-        Optimizer(g, self.executor._engine).optimize([])
+        RuntimeOptimizer(g, self.executor._engine).optimize([], False)
         self.assertTrue(any(n.op.__class__.__name__ == 'TensorCpFuseChunk' for n in g))
 
         c = next(n for n in g if n.op.__class__.__name__ == 'TensorCpFuseChunk')